+++ src/CSharpApp.Core/Interfaces/IProductsService.cs
@@ -5,6 +5,6 @@ namespace CSharpApp.Core.Interfaces;
 public interface IProductsService
 {
     public Task<IReadOnlyCollection<ProductDetails>> GetAll();
-    public Task<ProductDetails?> GetById(long id);
+    public Task<ProductDetails?> GetById(int id);
     public Task<ProductCreationResult> Create(ProductCreation product);
 }
